need to ensure lock

+++ streams/src/main/java/org/apache/kafka/streams/processor/internals/StateManagerUtil.java
@@ -100,41 +100,44 @@ final class StateManagerUtil {
             throw new IllegalArgumentException("State store could not be wiped out during clean close");
         }
 
-        ProcessorStateException exception = null;
-
         final TaskId id = stateMgr.taskId();
-        log.trace("Closing state manager for {}", id);
+        log.trace("Closing state manager for {} task {}", taskType, id);
 
+        ProcessorStateException exception = null;
         try {
-            stateMgr.close();
+            if (stateDirectory.lock(id)) {
+                try {
+                    stateMgr.close();
 
-            if (wipeStateStore) {
-                // we can just delete the whole dir of the task, including the state store images and the checkpoint files,
-                // and then we write an empty checkpoint file indicating that the previous close is graceful and we just
-                // need to re-bootstrap the restoration from the beginning
-                Utils.delete(stateMgr.baseDir());
-            }
-        } catch (final ProcessorStateException e) {
-            exception = e;
-        } catch (final IOException e) {
-            throw new ProcessorStateException("Failed to wiping state stores for task " + id, e);
-        } finally {
-            try {
-                stateDirectory.unlock(id);
-            } catch (final IOException e) {
-                if (exception == null) {
-                    exception = new ProcessorStateException(
-                        String.format("%sFailed to release state dir lock", logPrefix), e);
+                    if (wipeStateStore) {
+                        // we can just delete the whole dir of the task, including the state store images and the checkpoint files,
+                        // and then we write an empty checkpoint file indicating that the previous close is graceful and we just
+                        // need to re-bootstrap the restoration from the beginning
+                        Utils.delete(stateMgr.baseDir());
+                    }
+                } catch (final ProcessorStateException e) {
+                    exception = e;
+                } catch (final IOException e) {
+                    throw new ProcessorStateException("Failed to wiping state stores for task " + id, e);
+                } finally {
+                    try {
+                        stateDirectory.unlock(id);
+                    } catch (final IOException e) {
+                        if (exception == null) {
+                            exception = new ProcessorStateException(String.format("%sFailed to release state dir lock", logPrefix), e);
+                        }
+                    }
                 }
             }
+        } catch (final IOException e) {
+            throw new StreamsException(
+                String.format("%sFatal error while trying to lock the state directory for task %s", logPrefix, id),
+                e
+            );
         }
 
         if (exception != null) {
-            if (closeClean) {
-                throw exception;
-            } else {
-                log.warn("Closing {} task {} uncleanly and swallows an exception", taskType, id, exception);
-            }
+            throw exception;
         }
     }
 }
